bugFix/162211463_TxNoteRefactor (#826)
* Fixed Flow types
Fixed import order
Extracted method to utils
Fixed reducer structure
Fixed method for null messages case

* Removed json parse to optimize speed

* Fixed Android boolean check error in EventDetails
